apps/datahub-llm-description/requirements.txt
acryl-datahub==0.13.2.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
avro==1.11.3
avro-gen3==0.7.12
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
click-spinner==0.1.10
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
python-dotenv==1.0.1
expandvars==0.12.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
ijson==3.3.0
Jinja2==3.1.4
joblib==1.4.2
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llama-index-core==0.10.44
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-ollama==0.1.2
llama-index-llms-ollama==0.1.5
llama-index-readers-file==0.1.25
llamaindex-py-client==0.1.19
MarkupSafe==2.1.5
marshmallow==3.21.3
minijinja==2.0.1
mixpanel==4.10.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.33.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
progressbar2==4.4.2
psutil==6.0.0
pydantic==2.7.3
pydantic_core==2.18.4
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
sentry-sdk==2.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
